# openapi/tutorhub.yaml
openapi: 3.1.0
info:
  title: TutorHub Core API
  version: 0.0.1
  description: |
    Contract for TutorHub V2. Authenticated business operations use the active
    workspace from the server-side session and the shared authorization policy.
    A 403 response means the actor is known but lacks permission. A 404 response
    also covers resources outside the active tenant or class scope so identifiers
    cannot be enumerated across security boundaries.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/status:
    get:
      operationId: getAPIStatus
      summary: Return versioned Core API status
      responses:
        "200":
          description: Core API is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIStatusResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/auth/login:
    get:
      operationId: beginLogin
      summary: Start an OIDC Authorization Code and PKCE flow
      parameters:
        - name: return_to
          in: query
          required: false
          description: Internal application path used after successful sign-in
          schema:
            type: string
            example: /app/home
      responses:
        "303":
          description: Redirect to the configured identity provider
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/auth/callback:
    get:
      operationId: completeLogin
      summary: Complete the OIDC callback and issue a TutorHub session
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "303":
          description: Session created and browser redirected to TutorHub Web
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/auth/csrf:
    get:
      operationId: rotateCSRFToken
      summary: Rotate and return the CSRF token bound to the current session
      security:
        - SessionCookie: []
      responses:
        "200":
          description: CSRF token rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current TutorHub session
      security:
        - SessionCookie: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Local session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/me:
    get:
      operationId: getCurrentUser
      summary: Return the authenticated user and tenant context
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current authenticated principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/me/profile:
    get:
      operationId: getProfile
      summary: Return the authenticated user's editable profile
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Editable user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    patch:
      operationId: updateProfile
      summary: Update editable fields on the authenticated user's profile
      security:
        - SessionCookie: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/me/identities:
    get:
      operationId: listIdentities
      summary: List external identities linked to the authenticated user
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Linked external identities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /api/v1/me/identities/link:
    post:
      operationId: beginIdentityLink
      summary: Begin a recent-authenticated external identity linking flow
      security:
        - SessionCookie: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Provider authorization URL for the one-time linking flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityLinkResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /api/v1/session/active-tenant:
    put:
      operationId: switchActiveTenant
      summary: Switch the current session to an active tenant membership
      security:
        - SessionCookie: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchActiveTenantRequest"
      responses:
        "200":
          description: Active workspace changed and session credentials rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /api/v1/classes/{class_id}/leave:
    post:
      operationId: leaveClass
      summary: Leave a class enrollment as the authenticated actor
      security:
        - SessionCookie: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrollment left, including an idempotent repeated leave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassEnrollment"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: tutorhub_session
      description: Opaque TutorHub BFF session cookie; secure deployments use the __Host- prefix.
  responses:
    ProblemResponse:
      description: Request failed with a structured problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnauthorizedResponse:
      description: Authentication is missing, expired, or invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ForbiddenResponse:
      description: The authenticated actor lacks permission in the active workspace
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundResponse:
      description: The resource is absent or outside the active tenant/class scope
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ConflictResponse:
      description: The request conflicts with the current resource state or version
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    OrganizationRole:
      type: string
      enum: [org_admin, teacher, student, guest]
...
    ClassEnrollmentRole:
      type: string
      description: Persisted enrollment roles; class owner remains implicit in classes.owner_user_id.
      enum: [co_teacher, teaching_assistant, student]
    Permission:
      type: string
      enum:
        - tenant.view
        - tenant.manage
        - tenant.manage_members
        - tenant.manage_features
        - class.create
        - class.update
        - class.archive
        - class.transfer_ownership
        - class.view
        - enrollment.manage
        - enrollment.leave
        - session.start
        - session.end
        - session.join
        - participant.admit
        - participant.remove
        - media.publish
        - chat.send
        - audit.view
...
    APIStatusResponse:
      type: object
      additionalProperties: false
      required: [status, service, version, environment, timestamp]
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
        version:
          type: string
        environment:
          type: string
        timestamp:
          type: string
          format: date-time
    CSRFResponse:
      type: object
      additionalProperties: false
      required: [csrf_token]
      properties:
        csrf_token:
          type: string
    LogoutResponse:
      type: object
      additionalProperties: false
      properties:
        logout_url:
          type: string
          format: uri
    User:
      type: object
      additionalProperties: false
      required: [id, email, display_name, locale, timezone]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        locale:
          type: string
        timezone:
          type: string
        avatar_object_key:
          type: string
    ProfileResponse:
      type: object
      additionalProperties: false
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 120
        locale:
          type: string
          enum: [vi, en]
        timezone:
          type: string
          minLength: 1
          maxLength: 100
        avatar_object_key:
          type:
            - string
            - "null"
          maxLength: 512
    ExternalIdentity:
      type: object
      additionalProperties: false
      required:
        - id
        - provider
        - email
        - email_verified
        - created_at
        - last_authenticated_at
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_authenticated_at:
          type: string
          format: date-time
    IdentityListResponse:
      type: object
      additionalProperties: false
      required: [identities]
      properties:
        identities:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdentity"
    IdentityLinkResponse:
      type: object
      additionalProperties: false
      required: [authorization_url]
      properties:
        authorization_url:
          type: string
          format: uri
    TenantStatus:
      type: string
      enum: [active, suspended, archived]
    TenantMembership:
      type: object
      additionalProperties: false
      required: [id, slug, name, role, status, version, is_active]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/OrganizationRole"
        status:
          $ref: "#/components/schemas/TenantStatus"
        version:
          type: integer
          format: int64
          minimum: 1
        is_active:
          type: boolean
...
    SwitchActiveTenantRequest:
      type: object
      additionalProperties: false
      required: [tenant_id]
      properties:
        tenant_id:
          type: string
          format: uuid
    MeResponse:
      type: object
      additionalProperties: false
      required: [user, active_tenant, memberships, permissions]
      properties:
        user:
          $ref: "#/components/schemas/User"
        active_tenant:
          oneOf:
            - $ref: "#/components/schemas/TenantMembership"
            - type: "null"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/TenantMembership"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
...
    ClassEnrollmentStatus:
      type: string
      enum: [invited, active, suspended, left, removed]
...
    ClassEnrollment:
      type: object
      additionalProperties: false
      required:
        - id
        - class_id
        - user_id
        - class_role
        - status
        - enrolled_by
        - joined_at
        - suspended_at
        - left_at
        - removed_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        class_role:
          $ref: "#/components/schemas/ClassEnrollmentRole"
        status:
          $ref: "#/components/schemas/ClassEnrollmentStatus"
        enrolled_by:
          type: string
          format: uuid
        joined_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        suspended_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        left_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        removed_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          description: Stable machine-readable failure code such as feature_disabled, quota_exceeded, or feature_control_unavailable.
          maxLength: 64
          pattern: "^[a-z][a-z0-9_]*$"
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string